cambiado el tipo basico de clase en la estructura datos

+++ P3/includes/datos.h
@@ -8,7 +8,7 @@ typedef struct {
 	int nclases;
 	int ndatos;
 	double** atributos;
-	int** clase;
+	double** clase;
 	
 }datos;
